requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography==41.0.7
pydantic[email]==2.5.0
email-validator==2.1.0

# Rate Limiting & Caching
slowapi==0.1.9
redis==5.0.1

# Configuration
python-decouple==3.8
pydantic-settings==2.1.0

# HTTP requests for SSO
requests==2.31.0
httpx==0.25.2

# Middleware & CORS
starlette==0.27.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0


# requirements.txt (add these to your existing file)
# AWS and Cognito
boto3==1.34.0
python-jose[cryptography]==3.3.0

# Database
alembic==1.12.1

# For UUID generation
uuid==1.30

# Production
gunicorn==21.2.0
